requirements.txt
# LangChain and LLM
langchain
PyPDF2
pypdf
vertexai 
langchain_google_genai

# envirements
python-dotenv

# Memory database
chromadb